+++ tests/test_pytest_restrict.py
@@ -53,10 +53,9 @@ def test_it_restricts_one_class(testdir):
     )
     out = testdir.runpytest("--restrict-types", "unittest.TestCase")
     assert out.ret > 0
-    out.stderr.fnmatch_lines(
+    out.stdout.fnmatch_lines(
         [
-            "ERROR: test_one.py::TestA::test_a does not inherit from allowed "
-            + "pytest-restrict bases (unittest.TestCase)"
+            "*Failed: Test is not a type allowed by --restrict-types."
         ]
     )
 
@@ -81,10 +80,9 @@ def test_it_restricts_one_function(testdir):
     )
     out = testdir.runpytest("--restrict-types", "unittest.TestCase")
     assert out.ret > 0
-    out.stderr.fnmatch_lines(
+    out.stdout.fnmatch_lines(
         [
-            "ERROR: test_one.py::test_a does not inherit from allowed "
-            + "pytest-restrict bases (unittest.TestCase)"
+            "*Failed: Test is not a type allowed by --restrict-types."
         ]
     )
 
@@ -137,10 +135,8 @@ def test_it_restricts_multiple_types_not_allowed(testdir):
     )
     out = testdir.runpytest("--restrict-types", "my_test_base.A,my_test_base.B")
     assert out.ret > 0
-    out.stderr.fnmatch_lines(
+    out.stdout.fnmatch_lines(
         [
-            "ERROR: test_one.py::MyTests::test_one does not inherit from "
-            + "allowed pytest-restrict bases "
-            + "(my_test_base.A,my_test_base.B)"
+            "*Failed: Test is not a type allowed by --restrict-types."
         ]
     )